saved before merge

+++ My project/Assets/Scripts/BlueWeaponStatus.cs
@@ -4,17 +4,17 @@ public class BlueWeaponStatus : MonoBehaviour
 {
     public new string name = "blueweapon";
     public int level = 1;
-    public double weaponDamage = 1;
-    public double bossDamage = 0;
-    public double denyDefence = 0;
+    public float weaponDamage = 1;
+    public float bossDamage = 0;
+    public float denyDefence = 0;
 
-    public double Attack()
+    public float Attack()
     {
-        double damage = 0;
+        float damage = 0;
         damage = weaponDamage * (1 + bossDamage);
         return damage;
     }
-    public void initialize(int level, int bossDamage, double denyDefence)
+    public void initialize(int level, int bossDamage, float denyDefence)
     {
         this.level = level;
         weaponDamage = level * 3;
